+++ src/main/java/de/samply/email/EmailTemplateType.java
@@ -3,5 +3,7 @@ package de.samply.email;
 public enum EmailTemplateType {
     INVITATION,
     NEW_PROJECT,
-    FINISHED_PROJECT
+    FINISHED_PROJECT,
+    PROJECT_BRIDGEHEAD_ACCEPTED,
+    PROJECT_BRIDGEHEAD_REJECTED
 }